refactor: improve and move LetterManager to userInterface folder
