+++ src/userInterface/ILettersManager.ts
