+++ src/userInterface/LetterManager.ts
@@ -1,10 +1,27 @@
 import { ILetterManagerInterface } from "./ILettersManager.js";
 
 export class LetterManager  implements ILetterManagerInterface{
+    #turn: number;
+    #position: number;
     #letter: string;
     
-    constructor() {
+    constructor() { 
+        this.#turn = 0;
+        this.#position = 0;
         this.#letter = "";
+    }
+
+    get position(): number {
+        return this.#position;
+    }
+    set position(newPosition: number) {
+        this.#position = newPosition;
+    }
+    get turn(): number {
+        return this.#turn;
+    }
+    set turn(newTurn: number) {
+        this.#turn = newTurn;
     }
 
     get letter(): string {
